[@svelteui/core]: rename Breacrumb to Breadcrumbs and add unit test
