--- packages/svelteui-core/src/components/Breadcrumb/BreadcrumbItem/index.ts
@@ -1,3 +1,0 @@
-export { default as BreadcrumbItem } from './BreadcrumbItem.svelte';
-export type { BreadcrumbItemStylesParams } from './BreadcrumbItem.styles';
-export type { BreadcrumbItemProps } from './BreadcrumbItem';
