--- packages/svelteui-core/src/components/Breadcrumb/index.ts
@@ -1,3 +1,0 @@
-export { Breadcrumb } from './combinator';
-export type { BreadcrumbProps } from './Breadcrumb';
-export type { BreadcrumbItemProps, BreadcrumbItemStylesParams } from './BreadcrumbItem';
