+++ packages/svelteui-core/src/components/Breadcrumbs/Breadcrumbs.test.ts
@@ -1,0 +1,8 @@
+import { describe } from 'vitest';
+import { itSupportsClassName } from '@svelteuidev/tests';
+
+import { default as Breadcrumbs } from './Breadcrumbs.svelte';
+
+describe('Breadcrumbs', () => {
+	itSupportsClassName(Breadcrumbs, { mounted: true });
+});
