+++ packages/svelteui-core/src/components/Breadcrumbs/BreadcrumbsItem/index.ts
@@ -1,0 +1,3 @@
+export { default as BreadcrumbsItem } from './BreadcrumbsItem.svelte';
+export type { BreadcrumbItemStylesParams } from './BreadcrumbsItem.styles';
+export type { BreadcrumbItemProps } from './BreadcrumbsItem';
