+++ packages/svelteui-core/src/components/Breadcrumbs/index.ts
@@ -1,0 +1,3 @@
+export { Breadcrumbs } from './combinator';
+export type { BreadcrumbProps } from './Breadcrumbs';
+export type { BreadcrumbItemProps, BreadcrumbItemStylesParams } from './BreadcrumbsItem';
